% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactionlists.R
\name{addCors}
\alias{addCors}
\title{Adding cluster-specific correlations to annotated list of interactions}
\usage{
addCors(omicdata, intlist, memb)
}
\arguments{
\item{omicdata}{a list of matrices orresponding to omics types: "M", "CN", "T", "P" and "PP"; at least one continuous type must be present}

\item{intlist}{annotated list of interactions produced by the function \link{annotateEdges}}

\item{memb}{vector of memberships learned by the function \link{bnclustOmics}}
}
\description{
This function adds cluster-specific correlations to the list of annotated interactions; see \link{annotateEdges}
}
