% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpfns.R
\name{adjustMixedDir}
\alias{adjustMixedDir}
\title{Adjusting the PDAG matrix to model constraints
This function can be used to adjust the adjacency matrix to model constraints, such as blacklist and background nodes}
\usage{
adjustMixedDir(adj, bnnames, blacklist)
}
\arguments{
\item{adj}{adjacency matrix representing a graph or posterior probabilities of all its edges}

\item{bnnames}{object of class bnNames}

\item{blacklist}{a square matrix with same dimensions as adj representing edges prohibited by the model}
}
\value{
returns a matrix where entries prohibited by the model or blacklist are 0 and equal to correspondnding values of adj otherwise
}
\description{
Adjusting the PDAG matrix to model constraints
This function can be used to adjust the adjacency matrix to model constraints, such as blacklist and background nodes
}
