% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penalization.R
\name{blInit}
\alias{blInit}
\title{Initializing blacklist}
\usage{
blInit(
  bnnames,
  bldiag = TRUE,
  intra = NULL,
  interXX = list(from = NULL, to = NULL),
  interXY = list(from = NULL, to = NULL)
)
}
\arguments{
\item{bnnames}{object of class bnInfo; see \link{bnInfo}}

\item{bldiag}{logical, defines if diagonal should be blacklisted, TRUE by default}

\item{intra}{(optional) a vector of characters defining omic types for which intra-type edges will be blacklisted}

\item{interXX}{(optional) a list containing two vectors of characters defining omic types between which same gene (X.type.from -> X.type.to) edges will be blacklisted}

\item{interXY}{(optional) a list containing two vectors of characters defining omic types between which different gene edges  (X.type.from -> Y.type.to) will be blacklisted}
}
\value{
returns a binary matrix where 1 defines prohibited edges and 0 defines allowed edges
}
\description{
This function can be used to initialize a blacklist matrix for bnclustOmics clustering
}
\author{
Polina Suter
}
