% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penalization.R
\name{blUpdate}
\alias{blUpdate}
\title{Updating  blacklist}
\usage{
blUpdate(blacklist, node1, node2)
}
\arguments{
\item{blacklist}{object of class 'blacklist'}

\item{node1}{name of omic variable from which the edge is prohibited}

\item{node2}{name of omic variable to which the edge is prohibited}
}
\value{
returns a binary matrix where 1 defines prohibited interactions and 0 defines allowed interactions
}
\description{
This function can be used to update a blacklist matrix by blacklisting an edge between a pair of variables
}
\author{
Polina Suter
}
