% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnclustOmics.R
\name{bnclustOmics}
\alias{bnclustOmics}
\title{Bayesian network based clustering of multi-omics data}
\usage{
bnclustOmics(
  omicdata,
  bnnames,
  blacklist = NULL,
  edgepmat = NULL,
  kclust = 2,
  chixi = 0.5,
  seed = 100,
  err = 1e-06,
  maxEM = 10,
  hardlim = 6,
  deltahl = 2,
  nit = 5,
  epmatrix = TRUE,
  plus1it = 4,
  startpoint = "mclustPCA",
  baseprob = 0.4,
  commonspace = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{omicdata}{a list of matrices orresponding to omics types: "M", "CN", "T", "P" and "PP"; at least one continuous type must be present}

\item{bnnames}{object of class 'bnInfo'; see constructor function \link{bnInfo}}

\item{blacklist}{adjacency matrix containing information about which edges will be blacklisted in structure search}

\item{edgepmat}{penalization matrix of the edges in structure learning}

\item{kclust}{the number of clusters (mixture components)}

\item{chixi}{prior pseudocounts used for computing parameters for binary nodes}

\item{seed}{integer number set for reproducibility}

\item{err}{convergence criteria}

\item{maxEM}{maximum number of outer EM iterations (structural search)}

\item{hardlim}{maximum number of parents per node when learning networks}

\item{deltahl}{additional number of parents when sampling from the common search space}

\item{nit}{number of internal iteration (of parameter estimation) in the EM}

\item{epmatrix}{(logical) indicates if the matrices containing posterior probabilities of single edges should be returned}

\item{plus1it}{maximum number of search space expansion iterations when performing structure search}

\item{startpoint}{defines which algorithm is used to define starting cluster memberships: possible values "random", "mclustPCA" and "mclust"}

\item{baseprob}{defines the base probability of cluster membership when "mclustPCA" or "mclust" used as starting point}

\item{commonspace}{(logical) defines if the sampling has to be performed from the common search space}

\item{verbose}{defines if the output messages should be printed}
}
\value{
object of class 'bnclustOmics'
}
\description{
Bayesian network-based clustering of multi-omics data.
}
\author{
Polina Suter, Jack Kuipers
}
