% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{bnInfo}
\alias{bnInfo}
\title{Constructing object of class bnInfo}
\usage{
bnInfo(omicdata, types, omics, mappings = NULL, attachtype = FALSE)
}
\arguments{
\item{omicdata}{a list of matrices containing data, rows are observations, columns are variables (the order should be as following binary->ordinal->continuous)}

\item{types}{a vector of characters equal in length to the number of provided omic matrices, "b" binary, "o" ordinal, "c" continuous}

\item{omics}{a vector of omic names, must be the same as names of elements in omicdata, otherwise names of omicdata will be overwritten}

\item{mappings}{mappings containing a gene symbol for each omic type,rownames have to contain column names of the parameter 'omicdata'; column "gene" must be present}

\item{attachtype}{when TRUE .O will be attached to each variable name, where O is omic type; when FALSE (default) .O is only attached to duplicated names}
}
\description{
This function constructs an object of class bnInfo which is needed for Bayesian network
based clustering; see function \link{bnclustOmics}
}
