% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadescription.R
\docType{data}
\name{bnres4}
\alias{bnres4}
\title{bnres3}
\format{
An object of class 'bnclustOmics'
}
\usage{
bnres4
}
\description{
An object of class 'bnclustOmics' containing the results of one run of the function
'bnclustOmics' with the parameter k=4. The object contains membership assignments, estimated MAP
graphs representing clusters as well as posterior probabilities of all edges for each cluster.
}
\keyword{datasets}
