% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustfns.R
\name{checkmembership}
\alias{checkmembership}
\title{Comparing estimated and ground truth membership}
\usage{
checkmembership(k, truememb, estmemb)
}
\arguments{
\item{k}{number of clusters}

\item{truememb}{ground truth labels}

\item{estmemb}{estimated labels}
}
\description{
This function compares ground truth cluster labels to estimated cluster labels.
}
