% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractors.R
\name{chooseK}
\alias{chooseK}
\title{Choosing the number of clusters}
\usage{
chooseK(bnlist, fun = c("AIC", "BIC", "likel"))
}
\arguments{
\item{bnlist}{list of objects of class 'bnclustOmics'}

\item{fun}{score function for choosing the optimal number of clusters; available options are 'AIC' or  'BIC'}
}
\value{
a list consisting of a vector of scores extracted from each object of class bnclustOmics and the optimal k
}
\description{
This function can be used for choosing the optimal number of clusters using AIC or BIC scores.
}
\examples{
bnlist<-list()

#bnlist[[k]]<-bnclustOmics(simdata,bnnames,maxEM=4, kclust=k,startpoint = "mclustPCA")
bnlist[[2]]<-bnres2
bnlist[[3]]<-bnres3
bnlist[[4]]<-bnres4

chooseK(bnlist,fun="BIC")
chooseK(bnlist,fun="AIC")
}
