% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractors.R
\name{clusters}
\alias{clusters}
\title{Extracting cluster memberships}
\usage{
clusters(x, consensus = FALSE)
}
\arguments{
\item{x}{object of class 'bnclustOmics'}

\item{consensus}{logical, indicates if consensus clusters will be extracted; FALSE by default}
}
\value{
a vector of length of the number of observations corresponding to cluster assignments obtained by bnclustOmics
}
\description{
This function extracts a vector with MAP cluster
memberships assignments from the 'bnclustOmics' object
}
\examples{
clusters(bnres3)
}
