% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphfns.R
\name{getModels}
\alias{getModels}
\title{Deriving consensus graphs}
\usage{
getModels(bnres, p)
}
\arguments{
\item{bnres}{object of class 'bnclustOmics'}

\item{p}{posterior probability threshold}
}
\value{
a list of adjacency matrices corresponding to consensus graphs representing discovered clusters
}
\description{
This function derives a consensus grpah for each cluster according to a specified posterior probability.
}
\author{
Polina Suter
}
