% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadescription.R
\docType{data}
\name{mappings}
\alias{mappings}
\title{mappings}
\format{
a list of data frames, whose names correspond to omics types. The row names of each data frame correspond to IDs used in the data. At least one column "gene" is needed to specify gene symbol corresponding to the ID.
}
\usage{
mappings
}
\description{
An example of mappings needed for constructing bnInfo objects; a list of data frames, one for each omics type.
}
\keyword{datasets}
