% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractors.R
\name{posteriors}
\alias{posteriors}
\title{Extracting edge posterior probabilities}
\usage{
posteriors(x)
}
\arguments{
\item{x}{object of class 'bnclustOmics'}
}
\value{
a list of matrices containing posterior probabilities of all edges
in the graphs discovered by bnclustOmics when the parameter 'epmatrix' was set to TRUE
}
\description{
This function extracts a list of matrices containing posterior probabilities of all edges
in the graphs discovered by bnclustOmics when the parameter 'epmatrix' was set to TRUE
}
\examples{
post<-posteriors(bnres4)
}
