% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustfns.R
\name{relabelSimulation}
\alias{relabelSimulation}
\title{Relabeling clusters}
\usage{
relabelSimulation(res, trueclusters)
}
\arguments{
\item{res}{object of class 'bnclustOmics'}

\item{trueclusters}{ground truth clustering}
}
\value{
object of class 'bnclustOmics'
}
\description{
When running simulations studies, discovered cluster labels may differ from the ground truth
cluster labels. This functions can be used to perform relabeling in order to compare
discovered graphs to the ground truth graphs correctly.
}
