% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadescription.R
\docType{data}
\name{simclusters}
\alias{simclusters}
\title{simclusters}
\format{
a vector of 90 integers
}
\usage{
simclusters
}
\description{
Vector containing true cluster assigments for data in the dataset "simdata"
}
\keyword{datasets}
