% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadescription.R
\docType{data}
\name{simdags}
\alias{simdags}
\title{simdags}
\format{
a list of three binary matrices, each of size 70x70
}
\usage{
simdags
}
\description{
A list of three matrices representing adjacency matrices of DAGs used to generate the simulated dataset
'simdata'. Each DAG consists of 20 binary (mutations) and 50 continuous nodes (Gaussian).
}
\keyword{datasets}
