% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadescription.R
\docType{data}
\name{simdata}
\alias{simdata}
\title{simdata}
\format{
a list of two matrices: 'M', 90 rows (samples) and 20 columns (mutations), 'T' 90 rows and 100 columns (gene expression)
}
\usage{
simdata
}
\description{
A list of two matrices containing simulated mutations and transcriptome data (normalized, transformed) for three clusters. The generative
model is the mixture of Bayesian networks (linear Gaussian model). The networks are stored in the dataset 'simdags'. Ground truth cluster assigments are
stored in the dataset 'simclusters'
}
\keyword{datasets}
