% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadescription.R
\docType{data}
\name{simint}
\alias{simint}
\title{simint}
\format{
a data frame with two columns "gene1" and "gene2" and 73 rows
}
\usage{
simint
}
\description{
A list of interactions derived from the networks stored in 'simdags' that were used to generate
the dataset 'simdata'
}
\keyword{datasets}
