% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadescription.R
\docType{data}
\name{stringint}
\alias{stringint}
\title{stringint}
\format{
a data frames that includes three columns, gene1, gene2 and score. Column score is optional and may be skipped when constructing prior.
}
\usage{
stringint
}
\description{
An example of interactions list used for constructing graphical prior (penalization and blacklist) matrices.
}
\keyword{datasets}
