% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadescription.R
\docType{data}
\name{toydata}
\alias{toydata}
\title{toydata}
\format{
A list of five matrices,  one for each omics type. Each matrix contains 45 rows corresponding
to patient samples. The genes and gene products are in the columns.
\itemize{
\item M mutations, binary, 20 columns
\item CN copy number changes, ordinal, 20 columns
\item T transcriptome, continuous (Gaussian), 20 columns
\item P proteome, continuous (Gaussian), 15 columns
\item PP phospho-proteome, continuous (Gaussian), 20 columns
}
}
\usage{
toydata
}
\description{
Toy dataset containing five omics matrices that can be used for testing purposes.
}
\keyword{datasets}
