% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustfns.R
\name{checkmembership}
\alias{checkmembership}
\title{Comparing estimated and ground truth membership}
\usage{
checkmembership(estmemb, truememb)
}
\arguments{
\item{estmemb}{estimated labels}

\item{truememb}{ground truth labels}
}
\value{
a list containing different measures of similarity between two different clusterings, including accuracy, adjusted Rand index and precision
}
\description{
This function compares similarity between two clusterings.
}
